Added requests.js to utils with helper functions for making requests. Updated config.js to parse the query for GET requests.
