+++ client/components/dashboard/main/main.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import auth, {GetCurrentUser} from '../../utils/auth.js';
+import {GetCurrentUser} from '../../utils/auth.js';
+import { getUserCrews, getUserTasks } from '../../utils/requests.js';
 
 export default class Main extends Component {
 
@@ -9,48 +10,6 @@ export default class Main extends Component {
       user: null,
       userCrews: null,
       userTasks: null,
-    }
-    // Sets the path for 'fetches'
-    let url = process.env.HOST || 'http://localhost:3000/';
-
-    // Returns all of current user's crews. Will be rendered in sidebar view.
-    this.getUserCrews = () => {
-      let path = url + 'user/crews';
-      let options = {
-        method: 'GET',
-        headers: new Headers({
-          'Content-Type': 'application/json'
-        })
-      };
-      return fetch(path, options)
-      // TODO: test the data format of these API requests
-      // .then((response) => {
-      //   return response.json();
-      // })
-      .then((data) => {
-        console.log(data);
-        return this.setState({userCrews: data});
-      }).catch((error) => console.log('ERROR', error));
-    }
-
-    // Returns all of current user's tasks. Will be rendered in crew view.
-    this.getUserTasks = () => {
-      let path = url + 'user/tasks';
-      let options = {
-        method: 'GET',
-        headers: new Headers({
-          'Content-Type': 'application/json'
-        })
-      };
-      return fetch(path, options)
-      // TODO: test the data format of these API requests
-      // .then((response) => {
-      //   return response.json();
-      // })
-      .then((data) => {
-        console.log(data);
-        return this.setState({userTasks: data});
-      }).catch((error) => console.log('ERROR', error));
     }
   }
